Issue #61 - overhauling the parser for better error messages. (#149)
* Issue #61 - overhauling the parser for better error messages.

The big change here is we move from Attoparsec to Megaparsec. The
latter is nearly as fast in general, but has much better error
handling support.

The code changes aren't actually very dramatic, as the two libraries
are fundamentally quite similar. Summarising what I've actually needed to alter:

 - The double-parse we were doing before - once to parse, once to
   capture line breaks - is no longer necessary.
 - There are many simple function name changes. For example
   `skipSpaces` becomes `spaces`, `notChar` becomes `anySingleBut`.
 - Instead of storing `Position` as an `Int`, it's now an alias for
   megaparsec's `SourcePos`, which has richer position information.
 - The existing test data is almost exactly the same, except where the
   line/column numbers are more accurate than before.
 - I've labelled most of the parsers we use, for better error messages.
 - I've added some syntax error test cases. Obviously you can't
   predict every mistake a user can make, but what's there makes it clear
   that the error messages are much more useful.

I've also tried to move towards this convention:

> https://hackage.haskell.org/package/megaparsec-7.0.5/docs/Text-Megaparsec-Char-Lexer.html
>
> We propose a convention where every lexeme parser assumes no spaces
> before the lexeme and consumes all spaces after the lexeme;

I've found it generally makes it easier to resolve ambiguity, and know
who is responsible for consuming what whitespace. I haven't used the
lexeme code in that module though. I think it would probably be a good
idea, but wanted to keep this commit as uncontroversial as I could. :-)

* Parser enhancements - multiple commas between field names are perfectly legal.

* fixup! Parser enhancements - multiple commas between field names are perfectly legal.