rebuild when the schema changes (#557)
* rebuild when the schema changes

It is important to always call qAddDependentFile before reading a file
in TH, as it allows allows GHC to decide whether to rebuild a module or
not. Without qAddDependentFile, GHC will not rebuild the module if the
file TH read has changed but the module file has not. With
qAddDependentFile, GHC will rebuild the module if either file changed.

* update the relevant changelogs

* add missing "Minor Change" header

to clarify that the change is backwards-compatible