Avoid generating the same type multiple times.
When generating the code for different queries that share the same
schema, morphreus tries to generae the same type and instances multiple
times, which does not work.

With this change, before generating a type or instance, it will lookup
if it was declared already and avoid doing so.

This also helps when you want to take over and declare some things
manualy and let morpheus do the rest.

Additionally, this add a `omitNulls` pass to the generated Aeson object.
When sending the object for a type with lots of nullables, this reduces
significantly the request size.